refactor: rename telemetry/logging to telemetry/logger to avoid conflicts with the standard library logging module
