+++ pyproject.toml
@@ -1,10 +1,10 @@
 [build-system]
-requires = ["setuptools>=61.0", "wheel"]
+requires = ["setuptools>=82.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
 name = "panoseti-grpc"
-version = "0.3.4"
+version = "0.3.5"
 description = "gRPC for the PANOSETI project."
 readme = "README.md"
 requires-python = ">=3.9"
@@ -47,7 +47,7 @@ dependencies = [
     "psutil",
     "watchfiles",
     "aiofiles",
-#    "pygnssutils",
+    "pygnssutils",
     "pydantic>=2.12",
     "tomli",
     "requests",
